[Xposed] Android recovery mode (as update-binary)
Make the executable usable as a stand-alone "update-binary". When a ZIP archive
is flashed, the recovery extracts the file "META-INF/com/google/android/update-binary"
and runs it with three parameters (API version, file descriptor and zip file path).

The new logic detects such calls and does the following:
- Create a temporary directory.
- Extract the .zip file to it.
- Redirect all output to the file descriptor (with "ui_print" prefix).
- Run the file "META-INF/com/google/android/flash-script.sh".
- Remove the temporary directory.

This makes it possible to use the BusyBox executable as "update-binary" for flashable
ZIP archives. Within "flash-script.sh", the builtin applets are preferred, which
guarantees that commands work as expected (instead of having to rely on the target
device). This includes the initial "unzip" command.